Modularised openai.py, moved files in src/ to src/backend
Made encoding of base64 images handled seperatly in the openai.py file, and made the main.py handle the encoding of base64 images seperately. Moved src files into src/backend for more modularity and for ease of reading when dealing with future apis and user interfaces related to the project.